Split block.py 7/8: extract task-block family → models/task_blocks.py
Part 7/8. task_blocks.py (BaseTaskBlock + TaskBlock, ActionBlock, NavigationBlock, ExtractionBlock,
LoginBlock, ValidationBlock, HumanInteractionBlock, FileDownloadBlock, UrlBlock) imports Block and
helpers from block_base only — fully cycle-free (task_blocks-first import works). Absorbs SKY-11658's
task-block extraction; its block_base creation is already provided by the 1/8 foundation. Task-block
test patch targets repointed to task_blocks.
